Enable new Integration Test Framework in CI (#6087)
* Enable new Integration Test Framework in CI

This change enables the new tests to run in CI. In this process we found a few issues with both RNW and the framework. For the framework, we fixup an issue where we accidentally run code on the wrong thread (and add thread checks), add some clearer error messages, and add a specific error message if the instance fails to load instead of timing out.

Apart from builtin measures to show signal for when things go wrong, we add a utility to screenshot the CI desktop and run it before and after tests. We also upload crash dumps, and CI built symbols on failure for cases where we see native crashes.

- **Debug vs Release:** We run the debug build of the app to help catch extra assertsions and build a binary who should be easier to debug if failures show up. Things are quick enough that there aren't performance implications in doing this, but it might be worth running in release as well in the future for more coverage.
- **New job vs combination with existing**: This adds enough build time that it makes sense to be its own slice, especially if we see any instability that would cause needing to trigger multiple times.
- **Packaged vs Bundler:** We don't get accurate source maps when bundling. Using a packager helps with that instead,
- **Inline CI scripts vs separate:** The CI steps for integration tests aren't representative of the normal process, so there isn't much use of separate scripts. We do  more inline with the YAML instead for clarity.

* The crash happens on YellowBox

+++ packages/IntegrationTest/tests/SampleTests.tsx
@@ -28,6 +28,12 @@ functionTest.skip('FailingExample', () => {
  * Test methods can also be async
  */
 functionTest('AsyncNativeModuleExample', async () => {
+  // Using AsyncStorage will YellowBox, which crashes on pre-1903 (including
+  // CI) trying to load an image (#6085)
+  if (Platform.Version < 8) {
+    return;
+  }
+
   await AsyncStorage.clear();
   await AsyncStorage.setItem('foo', 'bar');
